feat(dev): CTL-1025 Phase 1 — canonical isTypingTarget + surface-constants refactor
- typing-target.ts: single TypingTargetLike guard covering INPUT/TEXTAREA/SELECT/contentEditable
- surface-constants.ts: pure (React-/router-free) extraction of Surface type + SURFACE_CHORD/LABEL/BREADCRUMB
- surface.ts: re-exports from both new modules; gains isTypingTarget re-export (adds SELECT coverage)
- key-nav.ts: classifyKey second param migrated from string tagName → TypingTargetLike structural shape
- use-keyboard-nav.ts: feeds document.activeElement as TypingTargetLike to classifyKey
- __tests__/key-nav.test.ts: migrated string-based isTypingTarget/classifyKey calls to { tagName } form
